requirements.txt
-e .
# Library dependencies for the python code.  You need to install these with
# `pip install -r requirements.txt` before you can run this.

click
shell
munch
ruamel.yaml
yapf
pylint